requirements.txt
beautifulsoup4==4.4.1
Django==1.9.2
djangorestframework==3.3.3
gunicorn==19.4.5
Markdown==2.6.6
Pillow==3.1.0
pitchfork==0.1.3
virtualenv==15.0.1
